fix: cast relationship count to integer to avoid bigint string
PostgreSQL COUNT(*) returns bigint, which the pg driver serializes as
a string. The frontend checks typeof === 'number' and falls back to 0.
Cast to ::integer so the value arrives as a JS number.

+++ packages/entity-engine/entity.service.ts
@@ -167,7 +167,7 @@ export class EntityService {
       const targetTableName = rel.targetEntity;
       const key = `${rel.name}Count`;
 
-      counts[key] = sql`(SELECT COUNT(*) FROM ${sql.raw(`"${targetTableName}"`)} WHERE ${sql.raw(`"${fkColumn}"`)} = ${sql.raw(`"${thisTableName}"."id"`)} AND "deleted_at" IS NULL AND ${tenantCondition()})`;
+      counts[key] = sql`(SELECT COUNT(*)::integer FROM ${sql.raw(`"${targetTableName}"`)} WHERE ${sql.raw(`"${fkColumn}"`)} = ${sql.raw(`"${thisTableName}"."id"`)} AND "deleted_at" IS NULL AND ${tenantCondition()})`;
     }
 
     return counts;
